subject update delete done

+++ iem-nodejs/routes/superAdmin.js
@@ -28,7 +28,7 @@ router.get('/add-user', function(req, res, next) {
     if(!validateAdmin(req, res)){
         return;
     }
-    res.render('superAdmin/add-user', {err: ''});
+    res.render('superAdmin/add-user', {err: '', success: ''});
 });
 
 /* POST ADD USER */
@@ -39,14 +39,30 @@ router.post('/add-user', function(req, res, next) {
     var u_id = req.body.u_id;
     var email = req.body.email;
     var log_as = req.body.log_as;
-    if(u_id === '' || email === '' || log_as !== 'stu' || log_as !== 'tch' ||
+    if(u_id === '' || email === '' ||
     u_id === undefined || email === undefined || log_as === undefined){
         var err = 'All fields are mandatory';
         res.render('superAdmin/add-user', {err: err});
         return;
     }
-    insertQuery(u_id, email, log_as);
-    res.render('superAdmin/add-user', {err: ''});
+    db.query('select * from auth where u_id = \''+u_id+'\' OR email = \''+email+'\'', function (err, result) {
+        if(result.length > 0){
+            if(result[0].email === email && result[0].u_id === u_id){
+                var err = 'This User ID and Email Id is already taken.';
+            }
+            else if(result[0].email === email) {
+                var err = 'This Email Id is already taken.';
+            }
+            else if(result[0].u_id === u_id) {
+                var err = 'This User Id is already taken.';
+            }
+            res.render('superAdmin/add-user', {err: err, success: ''});
+        }
+        else{
+            insertQuery(u_id, email, log_as);
+            res.render('superAdmin/add-user', {err: '', success: 'User Added Successfully'});
+        }
+    });
 });
 
 /* Send OTP */
@@ -169,7 +185,7 @@ router.post('/students', function(req, res, next) {
             return;
         }
     }
-    else if(req.body.action === 'delete'){
+    else if(req.body.action === 'delete'  && req.body.u_roll !== '' && req.body.u_roll !== undefined){
         deleteStudent(req.body.u_roll);
     }
     var query = '';
@@ -223,6 +239,9 @@ router.post('/update-student/', function(req, res, next) {
     var contact = req.body.contact;
     var dept = req.body.dept;
     var add_year = req.body.add_year;
+    if(u_id === undefined){
+        u_id = '';
+    }
     db.query('select * from student_details where u_roll = ?', u_id, function (err, result) {
         if(result.length === 0){
             res.render('superAdmin/student_update', {details: result, err: 'This user is not registered'});
@@ -277,9 +296,9 @@ router.post('/update-student/', function(req, res, next) {
 
 /* GET TEACHERS DETAILS */
 router.get('/teachers', function(req, res, next) {
-    // if(!validateAdmin(req, res)){
-    //     return;
-    // }
+    if(!validateAdmin(req, res)){
+        return;
+    }
     db.query('Select * from teacher_details', function (err, result) {
         res.render('superAdmin/teacher-table', {teacher: result});
     });
@@ -287,15 +306,15 @@ router.get('/teachers', function(req, res, next) {
 
 /* POST TEACHERS DETAILS */
 router.post('/teachers', function(req, res, next) {
-    // if(!validateAdmin(req, res)){
-    //     return;
-    // }
+    if(!validateAdmin(req, res)){
+        return;
+    }
     if(req.body.action === 'update') {
         if (updateTeacher(req, res)) {
             return;
         }
     }
-    else if(req.body.action === 'delete'){
+    else if(req.body.action === 'delete' && req.body.tch_id !== '' && req.body.tch_id !== undefined){
         deleteTeacher(req.body.tch_id);
     }
     res.redirect('/superAdmin/teachers');
@@ -317,9 +336,9 @@ function deleteTeacher(tch_id) {
 
 /* POST TEACHER UPDATE PAGE */
 router.post('/update-teacher/', function(req, res, next) {
-    // if(!validateAdmin(req, res)){
-    //     return;
-    // }
+    if(!validateAdmin(req, res)){
+        return;
+    }
     var u_id = req.body.u_id;
     var email = req.body.email;
     var f_name = req.body.f_name;
@@ -328,6 +347,9 @@ router.post('/update-teacher/', function(req, res, next) {
     var facebook_id = req.body.facebook_id;
     var googleplus_id = req.body.googleplus_id;
     var designation = req.body.designation;
+    if(u_id === undefined){
+        u_id = '';
+    }
     db.query('select * from teacher_details where tch_id = ?', u_id, function (err, result) {
         if(result.length === 0){
             res.render('superAdmin/teacher_update', {details: result, err: 'This user is not registered'});
@@ -382,33 +404,179 @@ router.post('/update-teacher/', function(req, res, next) {
 });
 
 
-/* GET SUBJECT UPDATE PAGE */
-router.get('/add-subject', function(req, res, next) {
-    if(!validateAdmin(req, res)){
-        return;
-    }
-    res.render('superAdmin/add-subject');
-});
-
-/* GET SUBJECT UPDATE PAGE */
-router.get('/add-subject', function(req, res, next) {
-    if(!validateAdmin(req, res)){
-        return;
-    }
-    res.render('superAdmin/subject_add');
-});
 
 
 
 
+
+
+
+
+
+
+/* GET subject page */
 router.get('/subjects', function(req, res, next) {
-    if(!validateAdmin(req, res)){
-        return;
-    }
+    // if(!validateAdmin(req, res)){
+    //     return;
+    // }
     db.query('Select * from subjects', function (err, result) {
-        res.render('superAdmin/subject-table', {subjects: result});
+        var dept = '';
+        var sem = '';
+        res.render('superAdmin/subject-table', {subjects: result, dept: dept, sem: sem});
     });
 });
+
+/* POST subject page */
+router.post('/subjects', function(req, res, next) {
+    // if(!validateAdmin(req, res)){
+    //     return;
+    // }
+    var sem = req.body.sem;
+    var dept = req.body.dept;
+    if(req.body.action === 'update') {
+        if (updateSubject(req, res)) {
+            return;
+        }
+    }
+    else if(req.body.action === 'delete' && req.body.sub_code !== '' && req.body.sub_code !== undefined){
+        deleteSubject(req.body.sub_code);
+    }
+    var query = '';
+    if( sem !== undefined && dept !== undefined){
+        query = 'Select * from subjects where dept = \''+dept+'\' and sem_code = \''+sem+'\' order by \'sub_code\'';
+    }
+    else if(sem !== undefined && dept === undefined){
+        query = 'Select * from subjects where sem_code = \''+sem+'\' order by \'sub_code\'';
+        dept = '';
+    }
+    else if(sem === undefined && dept !== undefined){
+        query = 'Select * from subjects where dept = \'' + dept +'\' order by \'sub_code\' ' ;
+        sem = '';
+    }
+    else{
+        query = 'Select * from subjects order by \'sub_code\' ' ;
+        dept = '';
+        sem = '';
+    }
+    db.query(query, function (err, result) {
+        res.render('superAdmin/subject-table', {subjects: result, dept: dept, sem: sem});
+    });
+});
+
+/* GET SUBJECT UPDATE PAGE */
+function updateSubject(req, res) {
+    // if(!validateAdmin(req, res)){
+    //     return;
+    // }
+    var sub_code = req.body.sub_code;
+    if(sub_code === '' || sub_code === undefined){
+        return 0;
+    }
+    db.query('select * from subjects where sub_code = ?', sub_code, function (err, result) {
+        res.render('superAdmin/update-subject', {err: '', subject: result, success: ''});
+    });
+    return 1;
+}
+
+function deleteSubject(sub_code) {
+    db.query('delete from subjects where sub_code = ?', sub_code, function (err, result) {});
+}
+
+
+/* GET SUBJECT UPDATE PAGE */
+router.post('/update-subject', function(req, res, next) {
+    // if(!validateAdmin(req, res)){
+    //     return;
+    // }
+    var sem_code = req.body.sem_code;
+    var sub_name = req.body.sub_name;
+    var sub_code = req.body.sub_code;
+    var dept = req.body.dept;
+    if(sub_code === undefined){
+        sub_code = '';
+    }
+    db.query('select * from subjects where sub_code = ?', req.body.old_sub_code,function (err, subject) {
+        if(sem_code === '' || sub_code === '' || sub_name === '' || dept=== '' ||
+            sem_code === undefined || sub_code === undefined || sub_name === undefined || dept=== undefined){
+            res.render('superAdmin/update-subject', {err: 'All fields are mandatory', success: '', subject: subject});
+            return;
+        }
+        if (subject[0].sub_code !== sub_code) {
+            db.query('select * from subjects where sub_code = ?', sub_code,function (err, newsubject) {
+                if(newsubject.length > 0){
+                    res.render('superAdmin/update-subject', {err: 'This Subject Code is already present', success: '', subject: subject});
+                }
+                else{
+                    db.query('delete from subjects where sub_code = ?', req.body.old_sub_code, function (err, result) {
+                        var arr = [sem_code, sub_code, sub_name, dept];
+                        db.query('Insert into subjects(sem_code, sub_code, sub_name, dept) values (?,?,?,?)', arr,function (err, result) {
+                            // res.render('superAdmin/update-subject', {err: '', success: 'Subject '+sub_code+' was successfully added', subject: subject});
+                            res.redirect('/superAdmin/subjects')
+                        });
+                    });
+                }
+            });
+        }
+        else{
+            db.query('delete from subjects where sub_code = ?', sub_code, function (err, result) {
+                var arr = [sem_code, sub_code, sub_name, dept];
+                db.query('Insert into subjects(sem_code, sub_code, sub_name, dept) values (?,?,?,?)', arr,function (err, result) {
+                    // res.render('superAdmin/update-subject', {err: '', success: 'Subject '+sub_code+' was successfully added', subject: subject});
+                    res.redirect('/superAdmin/subjects')
+                });
+            });
+        }
+    });
+});
+
+
+/* GET SUBJECT ADD PAGE */
+router.get('/add-subject', function(req, res, next) {
+    // if(!validateAdmin(req, res)){
+    //     return;
+    // }
+
+    res.render('superAdmin/add-subject', {err: '', success: ''});
+});
+
+
+/* GET SUBJECT ADD PAGE */
+router.post('/add-subject', function(req, res, next) {
+    // if(!validateAdmin(req, res)){
+    //     return;
+    // }
+    var sem_code = req.body.sem_code;
+    var sub_name = req.body.sub_name;
+    var sub_code = req.body.sub_code;
+    var dept = req.body.dept;
+    if(sem_code === '' || sub_code === '' || sub_name === '' || dept=== '' ||
+        sem_code === undefined || sub_code === undefined || sub_name === undefined || dept=== undefined){
+        res.render('superAdmin/add-subject', {err: 'All fields are mandatory', success: ''});
+        return;
+    }
+    db.query('select * from subjects where sub_code = ?', sub_code,function (err, result) {
+        if(result.length > 0){
+            res.render('superAdmin/add-subject', {err: 'This Subject Code is already present', success: ''});
+            return;
+        }
+        var arr = [sem_code, sub_code, sub_name, dept];
+        db.query('Insert into subjects(sem_code, sub_code, sub_name, dept) values (?,?,?,?)', arr,function (err, result) {
+            res.render('superAdmin/add-subject', {err: '', success: 'Subject '+sub_code+' was successfully added'});
+        });
+    });
+});
+
+
+
+
+
+
+
+
+
+
+
+
 
 
 /* GET superAdmin login page */
